+++ internal/cli/command/port.go
@@ -17,6 +17,7 @@ type PortService interface {
 type PortOptions struct {
 	Format   func() string
 	NoHeader func() bool
+	Color    func() bool
 }
 
 func NewPortCmd(s PortService, o PortOptions) *cobra.Command {
@@ -31,7 +32,7 @@ func NewPortCmd(s PortService, o PortOptions) *cobra.Command {
 		if err != nil {
 			return fmt.Errorf("collecting ports: %w", err)
 		}
-		r, err := render.New(o.Format(), render.WithNoHeader(o.NoHeader()))
+		r, err := render.New(o.Format(), render.WithNoHeader(o.NoHeader()), render.WithColor(o.Color()))
 		if err != nil {
 			return err
 		}